# docs/swagger.yaml
basePath: /
definitions:
...
  dto.SendContactRequest:
    properties:
      contact:
        items:
          $ref: '#/definitions/dto.SendContactItem'
        minItems: 1
        type: array
      delay:
        maximum: 300000
        minimum: 0
        type: integer
      mentioned:
        items:
          type: string
        type: array
      mentionsEveryOne:
        type: boolean
      number:
        type: string
      quoted:
        $ref: '#/definitions/dto.MessageRequestQuoted'
    required:
    - contact
    - number
    type: object
  dto.SendContactResponse:
    properties:
      instanceId:
        type: string
      key:
        $ref: '#/definitions/dto.MessageResponseKey'
      messageTimestamp:
        type: integer
      messageType:
        type: string
      status:
        type: string
    type: object
...
  utils.AuthenticationErrorResponse:
    properties:
      message:
        example: Unauthorized
        type: string
    type: object
  utils.HTTPErrorResponse:
    properties:
      error:
        description: |-
          Error is a legacy opaque value. Go's JSON encoder does not serialize an
          error message here, so clients should use Message and ErrorMessage.
        type: object
      errorMessage:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Verbeux AI
    url: https://github.com/verbeux-ai/whatsmiau
  description: Authenticated WhatsApp API compatible with Evolution API. The specification
    covers the complete application route surface, including canonical routes, compatibility
    aliases, calls and documentation delivery.
  license:
    name: MIT
  title: WhatsMiau API
  version: 0.3.2
paths:
...
  /v1/instance/{instance}/message/contact:
    post:
      consumes:
      - application/json
      description: Sends contact cards (vCard) to a WhatsApp number
      operationId: postV1InstanceByInstanceMessageContact
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Contact parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send one or more contacts
      tags:
      - Message
...
  /v1/instance/{instance}/message/location:
    post:
      consumes:
      - application/json
      description: Sends a geographic location to a WhatsApp number
      operationId: postV1InstanceByInstanceMessageLocation
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Location parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendLocationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a location message
      tags:
      - Message
  /v1/instance/{instance}/message/poll:
    post:
      consumes:
      - application/json
      description: Sends a multiple-choice poll to a WhatsApp number
      operationId: postV1InstanceByInstanceMessagePoll
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Poll parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendPollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendPollResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a poll
      tags:
      - Message
  /v1/instance/{instance}/message/status:
    post:
      consumes:
      - application/json
      description: Sends a status (text/image/audio/video) to status@broadcast
      operationId: postV1InstanceByInstanceMessageStatus
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Status parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a status broadcast
      tags:
      - Message
  /v1/instance/{instance}/message/sticker:
    post:
      consumes:
      - application/json
      description: Sends a WebP sticker to a WhatsApp number
      operationId: postV1InstanceByInstanceMessageSticker
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Sticker parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendStickerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendStickerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a sticker
      tags:
      - Message
...
  /v1/instance/{instance}/message/video:
    post:
      consumes:
      - application/json
      description: Sends a video file by URL to a WhatsApp number
      operationId: postV1InstanceByInstanceMessageVideo
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Video parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendDocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendDocumentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a video
      tags:
      - Message
...
  /v1/message/sendContact/{instance}:
    post:
      consumes:
      - application/json
      description: Sends contact cards (vCard) to a WhatsApp number
      operationId: postV1MessageSendContactByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Contact parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send one or more contacts
      tags:
      - Message
...
  /v1/message/sendLocation/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a geographic location to a WhatsApp number
      operationId: postV1MessageSendLocationByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Location parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendLocationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a location message
      tags:
      - Message
...
  /v1/message/sendPoll/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a multiple-choice poll to a WhatsApp number
      operationId: postV1MessageSendPollByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Poll parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendPollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendPollResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a poll
      tags:
      - Message
  /v1/message/sendPtv/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a round/circle video note to a WhatsApp number
      operationId: postV1MessageSendPtvByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: PTV parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendPtvRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendPtvResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a round video note (PTV)
      tags:
      - Message
...
  /v1/message/sendStatus/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a status (text/image/audio/video) to status@broadcast
      operationId: postV1MessageSendStatusByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Status parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a status broadcast
      tags:
      - Message
  /v1/message/sendSticker/{instance}:
    post:
      consumes:
      - application/json
      description: Sends a WebP sticker to a WhatsApp number
      operationId: postV1MessageSendStickerByInstance
      parameters:
      - description: Instance ID
        in: path
        name: instance
        required: true
        type: string
      - description: Sticker parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SendStickerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SendStickerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.AuthenticationErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send a sticker
      tags:
      - Message
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"